refactor(servers): merge metrics page into stats page and clean up unused imports
Remove standalone metrics page and integrate its functionality into the stats page
Clean up unused BarChart4 import from layout
